fix: update render logic
