+++ package.json
@@ -22,7 +22,7 @@
     "jsonresume-theme-short": "^0.1.7",
     "jsonresume-theme-spartan": "^0.3.0",
     "jsonresume-theme-stackoverflow": "^1.4.0",
-    "puppeteer-core": "^7.0.1"
+    "puppeteer-core": "^5.5.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^11.0.0",
@@ -38,7 +38,7 @@
     "lint-staged": "^10.5.3",
     "prettier": "^2.2.1",
     "pretty-quick": "^3.1.0",
-    "puppeteer": "^7.0.1",
+    "puppeteer": "^5.5.0",
     "vercel": "^21.2.3"
   },
   "husky": {
